Vite Typescript React Template
