+++ Algo-Visualzer/src/App.tsx
@@ -1,0 +1,6 @@
+import Message from './Message';
+
+function App() {
+  return <div><Message/></div>
+}
+export default App
