+++ Algo-Visualzer/src/Message.tsx
@@ -1,0 +1,4 @@
+function Message(){
+    return <h1>Hello World</h1> 
+}
+export default Message
