--- CSS_grid_test/index.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link href="styles.css" rel="stylesheet" type="text/css">
-</head>
-
-<body>
-        
-    <nav id="Menu">
-        <p>Algorithm Visualizer</p>
-        <ul>
-           <li><div>Algorithms ▼</div>
-               <ul>
-                   <div id="BFSButton"><a>BFS</a></div>
-                   <div id="DFSButton"><a>DFS</a></div>
-                   <div id="DijkstraButton"><a>Dijkstra's</a></div>
-               </ul> 
-            </li>
-           <li><div>Mazes ▼</div>
-                <ul>
-                   <div id="random"><a>Random</a></div>
-                   <div id="amongus"><a>Amongusඞ</a></div>
-                   <div id="america"><a>America</a></div>
-               </ul> 
-            </li>
-           <li><div id="StartButton">Visualize BFS!</div></li> 
-           <li><div id="Clear">Clear Board</div></li> 
-           <li><div id="ClearWalls">Clear Walls</div></li>
-           <li><div>Speed ▼</div>
-                <ul>
-                   <div id='slowButton'><a>Slow</a></div>
-                   <div id='mediumButton'><a>Medium</a></div>
-                   <div id='fastButton'><a>Fast</a></div>
-               </ul> 
-            </li>
-        </ul>
-    </nav>
-
-    <nav id="Tools">
-        <ul>
-            <li>
-                <div>
-                    <p id='startTool'></p>
-                    <p>Start Node</p>
-                </div>
-            </li>
-            <li>
-                <div>
-                    <p id='endTool'></p>
-                    <p>End Node</p>
-                </div>
-            </li>
-        </ul>
-    </nav>
-    <div class="grid" id='grid'></div>
-
-        <script src="Node.js"></script>
-        <script id="NodeFunctions" src="NodeFunctions.js"></script>
-        <script src="Algorithms/DFS.js"></script>
-        <script src="Algorithms/BFS.js"></script>
-        <script src="Algorithms/Dijkstra's.js"></script>
-        <script src="myscripts.js"></script>
-        
-    </div>
-    
-</body>
-</html>
