--- CSS_grid_test/styles.css
@@ -1,263 +1,0 @@
-
-    *{
-        margin: 0;
-        padding: 0;
-    }
-
-    #Menu{
-        width: 100%;
-        height: 80px;
-        background-color: #001f3f;
-    }
-
-    #Menu p{
-        cursor: pointer;
-        font-family: arial;
-        color: #FFFFFF;
-        font-size: 22px;
-        padding: 24px 14px;
-        float: left;
-    }
-
-    #Menu ul{
-        cursor: pointer;
-        float: left;
-        padding: 0px 50px;
-    }
-
-    #Menu ul li{
-        float: left;
-        list-style: none;
-        position: relative;
-        padding: 0px 10px;
-    }
-
-    #Menu ul li div{
-        font-family: arial;
-        color: #FFFFFF;
-        font-size: 22px;
-        padding: 24px 14px;
-        text-decoration: none;
-    }
-
-    #Menu ul li ul{
-        display: none;
-        position: absolute;
-        border-radius: 0px 0px 4px 4px;
-        background-color: #001f3f;
-    }
-
-    #Menu ul li:hover ul{
-        display: block;
-    }
-
-    #Menu ul li ul div{
-        padding: 10px 14px;
-    }
-
-    #Menu ul li ul div{
-        padding: 10px 14px;
-    }
-
-    #Menu ul li ul div:hover{
-        background-color: #111111;
-    }
-
-    #Menu ul li ul div a{
-        font-size: 18px
-    }
-
-    #Tools{
-        padding: 0;
-        width: 100%;
-        height: 100px;
-    }
-
-    #StartButton{
-        background-color: #FF4136;
-        margin-top: 10px;
-        height: 10px;
-    }
-
-    #Tools {
-        margin: 0;
-        padding: 0;
-        
-    }
-
-    #Tools ul li div{
-        margin-left: 500px;
-    }
-
-    #Tools ul li div p {
-      margin: 22px 10px;
-      float: left;
-    }
-
-    #startTool{
-        width: 25px;
-        height: 25px;
-        outline: 1px solid Lightgray;
-        background-color: green;
-        float: left;
-        margin: 20px 10px;
-    }
-
-    #endTool{
-        width: 25px;
-        height: 25px;
-        outline: 1px solid Lightgray;
-        background-color: red;
-        float: left;
-        margin: 20px 10px;
-    }
-    
-    .node{
-        width: 25px;
-        height: 25px;
-        outline: 1px solid Lightgray;
-        vertical-align: top;
-        display: inline-flex;
-    }
-
-    .node:hover{
-        background-color: darkgray;
-    }
-
-    .nodeBeingSearchedBFS{
-        width: 25px;
-        height: 25px;
-        outline: 1px solid Lightgray;
-        display: inline-block;
-        vertical-align: top;
-        animation-name: BFS;
-        animation-duration: 500ms;
-        animation-iteration-count: 1;
-    }
-
-    .nodeBeingSearchedDFS{
-        width: 25px;
-        height: 25px;
-        outline: 1px solid Lightgray;
-        display: inline-block;
-        vertical-align: top;
-        animation-name: DFS;
-        animation-duration: 1000ms;
-        animation-iteration-count: 1;
-    }
-
-    .nodeBeingSearchedDijkstra{
-        width: 25px;
-        height: 25px;
-        outline: 1px solid Lightgray;
-        display: inline-block;
-        vertical-align: top;
-        animation-name: Dijkstra;
-        animation-duration: 500ms;
-        animation-iteration-count: 1;
-    }
-
-    .wall{
-        width: 25px;
-        height: 25px;
-        outline: 1px solid #000036;
-        background-color: #000036;
-        vertical-align: top;
-        display: inline-flex;
-        animation: wallanimation;
-        animation-duration: 500ms;
-    }
-
-    @keyframes wallanimation {
-        0% {
-            transform: scale(.2, .2);
-        }
-        10%{
-            transform: scale(.5, .5);
-        }
-        50%{
-            transform: scale(2,2z);
-        }
-        100%{
-            transform: scale(1, 1);
-        }
-    }
-
-    .wall:hover{
-        background-color: whitesmoke;
-    }
-
-    @keyframes BFS {
-        0% {
-            background-color: darkslateblue;
-            transform: scale(1, 1);
-            box-shadow: 10014px 0 0 -5px #9880ff;
-        }
-        49% {
-            background-color: whitesmoke;
-            transform: scale(1, 1.5);
-            
-        }
-        50% {
-            background-color: MediumSpringGreen;
-            transform: scale(1, 1.5);
-            box-shadow: 10014px 0 0 2px #9880ff;
-        }
-        100% {
-            background-color:LightSeaGreen;
-            transform: scale(1, 1);
-            box-shadow: 10014px 0 0 -5px #9880ff;
-        }     
-    }
-
-    @keyframes DFS {
-        0%{
-            transform: scale(1, .1);
-            background-color: #111111
-        }
-        1%{
-            transform: scale(1, .2);
-            background-color: #F012BE;
-        }
-        10%{
-            transform: scale(1, .3);
-            background-color: red;
-        }
-        25%{
-            transform: scale(1, .5);
-            background-color: orangered;
-        }
-        60%{
-            transform: scale(1, 2);
-            background-color: whitesmoke;
-        }
-    }
-
-    @keyframes Dijkstra{
-        0% {
-            background-color: #FFDC00;
-            transform: scale(1, 6);
-            box-shadow: 10014px 0 0 -5px #9880ff;
-        }
-        49% {
-            background-color: #F012BE;
-            transform: scale(1, 2);
-            
-        }
-        50% {
-            background-color: #FF4136;
-            transform: scale(1, 1.5);
-            box-shadow: 10014px 0 0 2px #9880ff;
-        }
-        100% {
-            background-color:#FFFFFF;
-            transform: scale(1, 1);
-            box-shadow: 10014px 0 0 -5px #9880ff;
-        }     
-    }
-
-    .grid{
-        position: absolute;
-        margin: 20px 50px 0px;
-        padding: 0;
-    }
